+++ Sources/Utils/GUIUtils/TypeFormatters/AngleType.cs
@@ -4,7 +4,7 @@
 
 using System;
 
-namespace KSPDev.GUIUtils {
+namespace KSPDev.GUIUtils.TypeFormatters {
 
 /// <summary>
 /// Localized message formatting class for a numeric value that represents an <i>angle</i>. The